Completed vulnscan beta
Now running and testing
